+++ src/main/java/uk/gov/hmcts/reform/prl/services/citizen/CaseService.java
@@ -4,14 +4,9 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.ObjectUtils;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.ccd.client.model.EventRequestData;
 import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
@@ -21,26 +16,17 @@ import uk.gov.hmcts.reform.prl.clients.ccd.CcdCoreCaseDataService;
 import uk.gov.hmcts.reform.prl.enums.CaseEvent;
 import uk.gov.hmcts.reform.prl.enums.PartyEnum;
 import uk.gov.hmcts.reform.prl.enums.YesOrNo;
-import uk.gov.hmcts.reform.prl.enums.caseflags.PartyRole;
 import uk.gov.hmcts.reform.prl.mapper.citizen.CaseDataMapper;
 import uk.gov.hmcts.reform.prl.models.Element;
 import uk.gov.hmcts.reform.prl.models.UpdateCaseData;
-import uk.gov.hmcts.reform.prl.models.caseflags.Flags;
-import uk.gov.hmcts.reform.prl.models.caseflags.flagdetails.FlagDetail;
-import uk.gov.hmcts.reform.prl.models.caseflags.request.CitizenPartyFlagsManageRequest;
-import uk.gov.hmcts.reform.prl.models.caseflags.request.CitizenPartyFlagsRequest;
-import uk.gov.hmcts.reform.prl.models.caseflags.request.CitizenPartyUpdatedFlagsRequest;
-import uk.gov.hmcts.reform.prl.models.caseflags.request.FlagDetailRequest;
 import uk.gov.hmcts.reform.prl.models.caseinvite.CaseInvite;
 import uk.gov.hmcts.reform.prl.models.complextypes.PartyDetails;
-import uk.gov.hmcts.reform.prl.models.complextypes.PartyDetailsMeta;
 import uk.gov.hmcts.reform.prl.models.complextypes.WithdrawApplication;
 import uk.gov.hmcts.reform.prl.models.complextypes.citizen.User;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseData;
 import uk.gov.hmcts.reform.prl.models.user.UserInfo;
 import uk.gov.hmcts.reform.prl.repositories.CaseRepository;
 import uk.gov.hmcts.reform.prl.services.SystemUserService;
-import uk.gov.hmcts.reform.prl.services.caseflags.PartyLevelCaseFlagsService;
 import uk.gov.hmcts.reform.prl.services.noticeofchange.NoticeOfChangePartiesService;
 import uk.gov.hmcts.reform.prl.utils.CaseUtils;
 
@@ -70,9 +56,9 @@ import static uk.gov.hmcts.reform.prl.enums.noticeofchange.SolicitorRole.Represe
 import static uk.gov.hmcts.reform.prl.enums.noticeofchange.SolicitorRole.Representing.CARESPONDENT;
 import static uk.gov.hmcts.reform.prl.enums.noticeofchange.SolicitorRole.Representing.DAAPPLICANT;
 import static uk.gov.hmcts.reform.prl.enums.noticeofchange.SolicitorRole.Representing.DARESPONDENT;
-import static uk.gov.hmcts.reform.prl.utils.CaseUtils.getPartyDetailsMeta;
 import static uk.gov.hmcts.reform.prl.utils.ElementUtils.element;
 import static uk.gov.hmcts.reform.prl.utils.ElementUtils.wrapElements;
+
 
 @Slf4j
 @Service
@@ -106,8 +92,6 @@ public class CaseService {
 
     private final NoticeOfChangePartiesService noticeOfChangePartiesService;
     private static final String INVALID_CLIENT = "Invalid Client";
-
-    private final PartyLevelCaseFlagsService partyLevelCaseFlagsService;
 
     public CaseDetails updateCase(CaseData caseData, String authToken, String s2sToken,
                                   String caseId, String eventId, String accessCode) throws JsonProcessingException {
@@ -408,440 +392,4 @@ public class CaseService {
         return caseRepository.updateCase(authToken, caseId, updatedCaseData, CaseEvent.CITIZEN_CASE_WITHDRAW);
     }
 
-    public Flags getPartyCaseFlags(String authToken, String caseId, String partyId) {
-        CaseDetails caseDetails = getCase(authToken, caseId);
-        CaseData caseData = CaseUtils.getCaseData(caseDetails, objectMapper);
-        Optional<PartyDetailsMeta> partyDetailsMeta = getPartyDetailsMeta(
-            partyId,
-            caseData.getCaseTypeOfApplication(),
-            caseData
-        );
-
-        if (partyDetailsMeta.isPresent()
-            && partyDetailsMeta.get().getPartyDetails() != null
-            && !StringUtils.isEmpty(partyDetailsMeta.get().getPartyDetails().getLabelForDynamicList())) {
-            Optional<String> partyExternalCaseFlagField = getPartyExternalCaseFlagField(
-                caseData.getCaseTypeOfApplication(),
-                partyDetailsMeta.get().getPartyType(),
-                partyDetailsMeta.get().getPartyIndex()
-            );
-
-            if (partyExternalCaseFlagField.isPresent()) {
-                return objectMapper.convertValue(
-                    caseDetails.getData().get(partyExternalCaseFlagField.get()),
-                    Flags.class
-                );
-            }
-        }
-
-        return null;
-    }
-
-    public ResponseEntity<Object> createCitizenReasonableAdjustmentsFlags(
-        String caseId, String eventId, String authToken, CitizenPartyFlagsRequest citizenPartyFlagsRequest) {
-        log.info("Inside createCitizenReasonableAdjustmentsFlags caseId {}", caseId);
-        log.info("Inside createCitizenReasonableAdjustmentsFlags eventId {}", eventId);
-        log.info(
-            "Inside createCitizenReasonableAdjustmentsFlags citizenPartyFlagsRequest {}",
-            citizenPartyFlagsRequest
-        );
-
-        if (StringUtils.isEmpty(citizenPartyFlagsRequest.getPartyIdamId()) || ObjectUtils.isEmpty(
-            citizenPartyFlagsRequest.getPartyExternalFlags())) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("bad request");
-        }
-
-        UserDetails userDetails = idamClient.getUserDetails(authToken);
-        CaseEvent caseEvent = CaseEvent.fromValue(eventId);
-        EventRequestData eventRequestData = coreCaseDataService.eventRequest(
-            caseEvent,
-            userDetails.getId()
-        );
-
-        StartEventResponse startEventResponse =
-            coreCaseDataService.startUpdate(
-                authToken,
-                eventRequestData,
-                caseId,
-                false
-            );
-
-        CaseData caseData = CaseUtils.getCaseData(startEventResponse.getCaseDetails(), objectMapper);
-        Optional<PartyDetailsMeta> partyDetailsMeta = getPartyDetailsMeta(
-            citizenPartyFlagsRequest.getPartyIdamId(),
-            caseData.getCaseTypeOfApplication(),
-            caseData
-        );
-
-        if (!partyDetailsMeta.isPresent()
-            || null == partyDetailsMeta.get().getPartyDetails()) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("party details not found");
-        }
-
-        Optional<String> partyExternalCaseFlagField = getPartyExternalCaseFlagField(
-            caseData.getCaseTypeOfApplication(),
-            partyDetailsMeta.get().getPartyType(),
-            partyDetailsMeta.get().getPartyIndex()
-        );
-
-        Map<String, Object> caseDataMap = startEventResponse.getCaseDetails().getData();
-        if (!partyExternalCaseFlagField.isPresent()
-            || !caseDataMap.containsKey(partyExternalCaseFlagField.get())
-            || ObjectUtils.isEmpty(caseDataMap.get(partyExternalCaseFlagField.get()))) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("party external flag details not found");
-        }
-
-        try {
-            log.info("partyExternalCaseFlagField ===>" + objectMapper.writeValueAsString(partyExternalCaseFlagField.get()));
-        } catch (JsonProcessingException e) {
-            log.info("error");
-        }
-
-        Flags flags = objectMapper.convertValue(
-            caseDataMap.get(partyExternalCaseFlagField.get()),
-            Flags.class
-        );
-        try {
-            log.info("Existing external Party flags  ===>" + objectMapper.writeValueAsString(flags));
-        } catch (JsonProcessingException e) {
-            log.info("error");
-        }
-        flags.getDetails().addAll(convertFlags(citizenPartyFlagsRequest.getPartyExternalFlags().getDetails()));
-
-        try {
-            log.info("Updated external Party flags  ===>" + objectMapper.writeValueAsString(flags));
-        } catch (JsonProcessingException e) {
-            log.info("error");
-        }
-        Map<String, Object> updatedCaseDataMap = new HashMap<>();
-        updatedCaseDataMap.put(partyExternalCaseFlagField.get(), flags);
-
-        CaseDataContent caseDataContent = coreCaseDataService.createCaseDataContent(
-            startEventResponse,
-            updatedCaseDataMap
-        );
-
-        try {
-            log.info("Case data content is  ===>" + objectMapper.writeValueAsString(caseDataContent));
-        } catch (JsonProcessingException e) {
-            log.info("error");
-        }
-
-        coreCaseDataService.submitUpdate(
-            authToken,
-            eventRequestData,
-            caseDataContent,
-            caseId,
-            false
-        );
-        return ResponseEntity.status(HttpStatus.OK).body("party flags updated");
-    }
-
-    public ResponseEntity<Object> manageCitizenReasonableAdjustmentsFlags(
-        String caseId, String eventId, String authToken, CitizenPartyFlagsManageRequest citizenPartyFlagsManageRequest) {
-        log.info("Inside createCitizenReasonableAdjustmentsFlags caseId {}", caseId);
-        log.info("Inside createCitizenReasonableAdjustmentsFlags eventId {}", eventId);
-        log.info(
-            "Inside createCitizenReasonableAdjustmentsFlags citizenPartyFlagsRequest {}",
-            citizenPartyFlagsManageRequest
-        );
-
-        if (StringUtils.isEmpty(citizenPartyFlagsManageRequest.getPartyIdamId()) || ObjectUtils.isEmpty(
-            citizenPartyFlagsManageRequest.getPartyExternalFlags())) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("bad request");
-        }
-
-        UserDetails userDetails = idamClient.getUserDetails(authToken);
-        CaseEvent caseEvent = CaseEvent.fromValue(eventId);
-        EventRequestData eventRequestData = coreCaseDataService.eventRequest(
-            caseEvent,
-            userDetails.getId()
-        );
-
-        StartEventResponse startEventResponse =
-            coreCaseDataService.startUpdate(
-                authToken,
-                eventRequestData,
-                caseId,
-                false
-            );
-
-        CaseData caseData = CaseUtils.getCaseData(startEventResponse.getCaseDetails(), objectMapper);
-        Optional<PartyDetailsMeta> partyDetailsMeta = getPartyDetailsMeta(
-            citizenPartyFlagsManageRequest.getPartyIdamId(),
-            caseData.getCaseTypeOfApplication(),
-            caseData
-        );
-
-        if (!partyDetailsMeta.isPresent()
-            || null == partyDetailsMeta.get().getPartyDetails()) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("party details not found");
-        }
-
-        Optional<String> partyExternalCaseFlagField = getPartyExternalCaseFlagField(
-            caseData.getCaseTypeOfApplication(),
-            partyDetailsMeta.get().getPartyType(),
-            partyDetailsMeta.get().getPartyIndex()
-        );
-
-        Map<String, Object> caseDataMap = startEventResponse.getCaseDetails().getData();
-        if (!partyExternalCaseFlagField.isPresent()
-            || !caseDataMap.containsKey(partyExternalCaseFlagField.get())
-            || ObjectUtils.isEmpty(caseDataMap.get(partyExternalCaseFlagField.get()))) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("party external flag details not found");
-        }
-
-        try {
-            log.info("partyExternalCaseFlagField ===>" + objectMapper.writeValueAsString(partyExternalCaseFlagField.get()));
-        } catch (JsonProcessingException e) {
-            log.info("error");
-        }
-
-        Flags flags = objectMapper.convertValue(
-            caseDataMap.get(partyExternalCaseFlagField.get()),
-            Flags.class
-        );
-        try {
-            log.info("Existing external Party flags  ===>" + objectMapper.writeValueAsString(flags));
-        } catch (JsonProcessingException e) {
-            log.info("error");
-        }
-
-        flags = flags.toBuilder()
-            .details(convertManageFlags(flags, citizenPartyFlagsManageRequest.getPartyExternalFlags().getDetails()))
-            .build();
-
-        try {
-            log.info("Updated external Party flags  ===>" + objectMapper.writeValueAsString(flags));
-        } catch (JsonProcessingException e) {
-            log.info("error");
-        }
-        Map<String, Object> updatedCaseDataMap = new HashMap<>();
-        updatedCaseDataMap.put(partyExternalCaseFlagField.get(), flags);
-
-        CaseDataContent caseDataContent = coreCaseDataService.createCaseDataContent(
-            startEventResponse,
-            updatedCaseDataMap
-        );
-
-        try {
-            log.info("Case data content is  ===>" + objectMapper.writeValueAsString(caseDataContent));
-        } catch (JsonProcessingException e) {
-            log.info("error");
-        }
-
-        coreCaseDataService.submitUpdate(
-            authToken,
-            eventRequestData,
-            caseDataContent,
-            caseId,
-            false
-        );
-        return ResponseEntity.status(HttpStatus.OK).body("party flags updated");
-    }
-
-    public ResponseEntity<Object> replaceCitizenReasonableAdjustmentsFlags(
-        String caseId, String eventId, String authToken, CitizenPartyUpdatedFlagsRequest citizenPartyUpdatedFlagsRequest) {
-        log.info("Inside replaceCitizenReasonableAdjustmentsFlags caseId {}", caseId);
-        log.info("Inside replaceCitizenReasonableAdjustmentsFlags eventId {}", eventId);
-        log.info(
-            "Inside replaceCitizenReasonableAdjustmentsFlags citizenPartyFlagsRequest {}",
-            citizenPartyUpdatedFlagsRequest
-        );
-
-        if (StringUtils.isEmpty(citizenPartyUpdatedFlagsRequest.getPartyIdamId()) || ObjectUtils.isEmpty(
-            citizenPartyUpdatedFlagsRequest.getPartyExternalFlags())) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("bad request");
-        }
-
-        UserDetails userDetails = idamClient.getUserDetails(authToken);
-        CaseEvent caseEvent = CaseEvent.fromValue(eventId);
-        EventRequestData eventRequestData = coreCaseDataService.eventRequest(
-            caseEvent,
-            userDetails.getId()
-        );
-
-        StartEventResponse startEventResponse =
-            coreCaseDataService.startUpdate(
-                authToken,
-                eventRequestData,
-                caseId,
-                false
-            );
-
-        CaseData caseData = CaseUtils.getCaseData(startEventResponse.getCaseDetails(), objectMapper);
-        Optional<PartyDetailsMeta> partyDetailsMeta = getPartyDetailsMeta(
-            citizenPartyUpdatedFlagsRequest.getPartyIdamId(),
-            caseData.getCaseTypeOfApplication(),
-            caseData
-        );
-
-        if (!partyDetailsMeta.isPresent()
-            || null == partyDetailsMeta.get().getPartyDetails()) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("party details not found");
-        }
-
-        Optional<String> partyExternalCaseFlagField = getPartyExternalCaseFlagField(
-            caseData.getCaseTypeOfApplication(),
-            partyDetailsMeta.get().getPartyType(),
-            partyDetailsMeta.get().getPartyIndex()
-        );
-
-        Map<String, Object> caseDataMap = startEventResponse.getCaseDetails().getData();
-        if (!partyExternalCaseFlagField.isPresent()
-            || !caseDataMap.containsKey(partyExternalCaseFlagField.get())
-            || ObjectUtils.isEmpty(caseDataMap.get(partyExternalCaseFlagField.get()))) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("party external flag details not found");
-        }
-
-        try {
-            log.info("partyExternalCaseFlagField ===>" + objectMapper.writeValueAsString(partyExternalCaseFlagField.get()));
-        } catch (JsonProcessingException e) {
-            log.info("error");
-        }
-
-        Flags flags = objectMapper.convertValue(
-            caseDataMap.get(partyExternalCaseFlagField.get()),
-            Flags.class
-        );
-        try {
-            log.info("Existing external Party flags  ===>" + objectMapper.writeValueAsString(flags));
-        } catch (JsonProcessingException e) {
-            log.info("error");
-        }
-        flags = flags.toBuilder()
-            .details(convertFlags1(citizenPartyUpdatedFlagsRequest.getPartyExternalFlags().getDetails()))
-            .build();
-
-        try {
-            log.info("Updated external Party flags  ===>" + objectMapper.writeValueAsString(flags));
-        } catch (JsonProcessingException e) {
-            log.info("error");
-        }
-        Map<String, Object> updatedCaseDataMap = new HashMap<>();
-        updatedCaseDataMap.put(partyExternalCaseFlagField.get(), flags);
-
-        CaseDataContent caseDataContent = coreCaseDataService.createCaseDataContent(
-            startEventResponse,
-            updatedCaseDataMap
-        );
-
-        try {
-            log.info("Case data content is  ===>" + objectMapper.writeValueAsString(caseDataContent));
-        } catch (JsonProcessingException e) {
-            log.info("error");
-        }
-
-        coreCaseDataService.submitUpdate(
-            authToken,
-            eventRequestData,
-            caseDataContent,
-            caseId,
-            false
-        );
-        return ResponseEntity.status(HttpStatus.OK).body("party flags updated");
-    }
-
-    private List<Element<FlagDetail>> convertFlags(List<FlagDetailRequest> details) {
-        List<Element<FlagDetail>> flagDetails = new ArrayList<>();
-
-        for (FlagDetailRequest detail : details) {
-            FlagDetail flagDetail = FlagDetail.builder().name(detail.getName())
-                .name_cy(detail.getName_cy())
-                .subTypeValue(detail.getSubTypeValue())
-                .subTypeValue_cy(detail.getSubTypeValue_cy())
-                .subTypeKey(detail.getSubTypeKey())
-                .otherDescription(detail.getOtherDescription())
-                .otherDescription_cy(detail.getOtherDescription_cy())
-                .flagComment(detail.getFlagComment())
-                .flagComment_cy(detail.getFlagComment_cy())
-                .flagUpdateComment(detail.getFlagUpdateComment())
-                .dateTimeCreated(detail.getDateTimeCreated())
-                .dateTimeModified(detail.getDateTimeModified())
-                .path(detail.getPath())
-                .hearingRelevant(detail.getHearingRelevant())
-                .flagCode(detail.getFlagCode())
-                .status(detail.getStatus())
-                .availableExternally(detail.getAvailableExternally())
-                .build();
-            flagDetails.add(element(flagDetail));
-        }
-
-        return flagDetails;
-    }
-
-    private List<Element<FlagDetail>> convertManageFlags(Flags flags, List<Element<FlagDetailRequest>> details) {
-        List<Element<FlagDetail>> flagDetails = flags.getDetails();
-
-        for (Element<FlagDetail> flagDetail : flagDetails) {
-            Optional<Element<FlagDetailRequest>> data
-                = details.stream().filter(x -> x.getId().equals(flagDetail.getId())).findFirst();
-
-            if (data.isPresent()) {
-
-                FlagDetail flagData = flagDetail.getValue().toBuilder().status(data.get().getValue().getStatus())
-                    .dateTimeModified(data.get().getValue().getDateTimeModified())
-                    .build();
-
-                int index = flagDetails.indexOf(flagDetail);
-                flagDetails.set(index, element(flagDetail.getId(), flagData));
-            }
-        }
-
-        return flagDetails;
-    }
-
-    private List<Element<FlagDetail>> convertFlags1(List<Element<FlagDetailRequest>> details) {
-        List<Element<FlagDetail>> flagDetails = new ArrayList<>();
-
-        for (Element<FlagDetailRequest> detail1 : details) {
-            FlagDetailRequest detail = detail1.getValue();
-            FlagDetail flagDetail = FlagDetail.builder().name(detail.getName())
-                .name_cy(detail.getName_cy())
-                .subTypeValue(detail.getSubTypeValue())
-                .subTypeValue_cy(detail.getSubTypeValue_cy())
-                .subTypeKey(detail.getSubTypeKey())
-                .otherDescription(detail.getOtherDescription())
-                .otherDescription_cy(detail.getOtherDescription_cy())
-                .flagComment(detail.getFlagComment())
-                .flagComment_cy(detail.getFlagComment_cy())
-                .flagUpdateComment(detail.getFlagUpdateComment())
-                .dateTimeCreated(detail.getDateTimeCreated())
-                .dateTimeModified(detail.getDateTimeModified())
-                .path(detail.getPath())
-                .hearingRelevant(detail.getHearingRelevant())
-                .flagCode(detail.getFlagCode())
-                .status(detail.getStatus())
-                .availableExternally(detail.getAvailableExternally())
-                .build();
-
-            flagDetails.add(element(detail1.getId(), flagDetail));
-        }
-
-        return flagDetails;
-    }
-
-    private Optional<String> getPartyExternalCaseFlagField(String caseType, PartyEnum partyType, Integer partyIndex) {
-
-        Optional<String> partyExternalCaseFlagField = Optional.empty();
-        boolean isC100Case = C100_CASE_TYPE.equalsIgnoreCase(caseType);
-
-        if (PartyEnum.applicant == partyType) {
-            partyExternalCaseFlagField
-                = Optional.ofNullable(partyLevelCaseFlagsService.getPartyCaseDataExternalField(
-                caseType,
-                isC100Case ? PartyRole.Representing.CAAPPLICANT : PartyRole.Representing.DAAPPLICANT,
-                partyIndex
-            ));
-        } else if (PartyEnum.respondent == partyType) {
-            partyExternalCaseFlagField
-                = Optional.ofNullable(partyLevelCaseFlagsService.getPartyCaseDataExternalField(
-                caseType,
-                isC100Case ? PartyRole.Representing.CARESPONDENT : PartyRole.Representing.DARESPONDENT,
-                partyIndex
-            ));
-        }
-
-        return partyExternalCaseFlagField;
-    }
 }
