+++ src/test/java/uk/gov/hmcts/reform/prl/controllers/citizen/CaseControllerTest.java
@@ -297,42 +297,42 @@ public class CaseControllerTest {
         );
     }
 
-    //    @Test
-    //    public void testCitizenRetrieveCases() {
-    //
-    //        List<CaseData> caseDataList = new ArrayList<>();
-    //
-    //
-    //        caseData = CaseData.builder()
-    //            .id(1234567891234567L)
-    //            .applicantCaseName("test")
-    //            .build();
-    //
-    //        caseDataList.add(CaseData.builder()
-    //                             .id(1234567891234567L)
-    //                             .applicantCaseName("test")
-    //                             .build());
-    //
-    //        when(authorisationService.authoriseService(any())).thenReturn(true);
-    //        when(authorisationService.authoriseUser(any())).thenReturn(true);
-    //
-    //        List<CaseDetails> caseDetails = new ArrayList<>();
-    //
-    //        Map<String, Object> stringObjectMap = caseData.toMap(new ObjectMapper());
-    //        caseDetails.add(CaseDetails.builder().id(
-    //            1234567891234567L).data(stringObjectMap).build());
-    //
-    //        String userId = "12345";
-    //        String role = "test role";
-    //
-    //        List<CaseData> caseDataList1 = new ArrayList<>();
-    //
-    //        when(objectMapper.convertValue(stringObjectMap, CaseData.class)).thenReturn(caseData);
-    //        when(caseService.retrieveCases(authToken, servAuthToken)).thenReturn(caseDataList);
-    //        caseDataList1 = caseController.retrieveCases(role, userId, authToken, servAuthToken);
-    //        assertNotNull(caseDataList1);
-    //
-    //    }
+    @Test
+    public void testCitizenRetrieveCases() {
+
+        List<CaseData> caseDataList = new ArrayList<>();
+
+
+        caseData = CaseData.builder()
+            .id(1234567891234567L)
+            .applicantCaseName("test")
+            .build();
+
+        caseDataList.add(CaseData.builder()
+                             .id(1234567891234567L)
+                             .applicantCaseName("test")
+                             .build());
+
+        when(authorisationService.authoriseService(any())).thenReturn(true);
+        when(authorisationService.authoriseUser(any())).thenReturn(true);
+
+        List<CaseDetails> caseDetails = new ArrayList<>();
+
+        Map<String, Object> stringObjectMap = caseData.toMap(new ObjectMapper());
+        caseDetails.add(CaseDetails.builder().id(
+            1234567891234567L).data(stringObjectMap).build());
+
+        String userId = "12345";
+        String role = "test role";
+
+        List<CaseData> caseDataList1 = new ArrayList<>();
+
+        when(objectMapper.convertValue(stringObjectMap, CaseData.class)).thenReturn(caseData);
+        when(caseService.retrieveCases(authToken, servAuthToken)).thenReturn(caseDataList);
+        caseDataList1 = caseController.retrieveCases(role, userId, authToken, servAuthToken);
+        assertNotNull(caseDataList1);
+
+    }
 
     @Test
     public void testCitizenLinkDefendantToClaim() {
@@ -507,100 +507,4 @@ public class CaseControllerTest {
 
         throw new RuntimeException("Invalid Client");
     }
-
-    //    @Test
-    //    public void testRetrieveRaFlags() {
-    //        String caseId = "1234567891234567L";
-    //        String partyId = "e3ceb507-0137-43a9-8bd3-85dd23720648";
-    //
-    //        Mockito.when(authorisationService.authoriseUser(authToken)).thenReturn(Boolean.TRUE);
-    //        Mockito.when(authorisationService.authoriseService(servAuthToken)).thenReturn(Boolean.TRUE);
-    //        Mockito.when(authTokenGenerator.generate()).thenReturn(servAuthToken);
-    //        Mockito.when(caseService.getPartyCaseFlags(authToken, caseId, partyId)).thenReturn(Flags.builder().roleOnCase(
-    //            "Respondent 1").partyName("Respondent").details(
-    //            Collections.emptyList()).build());
-    //
-    //        Flags flag = caseController.getCaseFlags(caseId, partyId, authToken, servAuthToken);
-    //        Assert.assertNotNull(flag);
-    //    }
-
-    //    @Test
-    //    public void testRetrieveRaFlagsWhenAuthFails() throws IOException {
-    //        expectedEx.expect(RuntimeException.class);
-    //        expectedEx.expectMessage("Invalid Client");
-    //
-    //
-    //        String caseId = "1234567891234567L";
-    //        String partyId = "e3ceb507-0137-43a9-8bd3-85dd23720648";
-    //
-    //        Mockito.when(authorisationService.authoriseUser(authToken)).thenReturn(Boolean.FALSE);
-    //        Mockito.when(authorisationService.authoriseService(servAuthToken)).thenReturn(Boolean.TRUE);
-    //
-    //        caseController.getCaseFlags(caseId, partyId, authToken, servAuthToken);
-    //
-    //        throw new RuntimeException("Invalid Client");
-    //    }
-
-    //    @Test
-    //    public void testUpdateCitizenRAflags() {
-    //        String caseId = "1234567891234567L";
-    //        String eventId = "c100RequestSupport";
-    //        String partyId = "e3ceb507-0137-43a9-8bd3-85dd23720648";
-    //        FlagDetailRequest flagDetailRequest = FlagDetailRequest.builder()
-    //            .name("Support filling in forms")
-    //            .name_cy("Cymorth i lenwi ffurflenni")
-    //            .hearingRelevant(YesOrNo.No)
-    //            .flagCode("RA0018")
-    //            .status("Requested")
-    //            .availableExternally(YesOrNo.Yes)
-    //            .build();
-    //        List<FlagDetailRequest> flagDetailsRequest = Collections.singletonList(flagDetailRequest);
-    //        CitizenPartyFlagsRequest partyRequestFlags = CitizenPartyFlagsRequest.builder()
-    //            .caseTypeOfApplication("C100")
-    //            .partyIdamId(partyId)
-    //            .partyExternalFlags(FlagsRequest.builder()
-    //                                    .details(flagDetailsRequest).build()).build();
-    //        Mockito.when(authorisationService.authoriseUser(authToken)).thenReturn(Boolean.TRUE);
-    //        Mockito.when(authorisationService.authoriseService(servAuthToken)).thenReturn(Boolean.TRUE);
-    //        Mockito.when(authTokenGenerator.generate()).thenReturn(servAuthToken);
-    //        Mockito.when(caseService.createCitizenReasonableAdjustmentsFlags(
-    //            caseId,
-    //            eventId,
-    //            authToken,
-    //            partyRequestFlags
-    //        )).thenReturn(ResponseEntity.status(HttpStatus.OK).body("party flags updated"));
-    //
-    //        ResponseEntity<Object> updateResponse = caseController.updateCitizenRAflags(
-    //        partyRequestFlags, eventId, caseId, authToken, servAuthToken);
-    //        assertThat(updateResponse.getStatusCode()).isEqualTo(HttpStatus.OK);
-    //    }
-
-    //    @Test
-    //    public void testUpdateCitizenRAflagsWhenAuthFails() {
-    //        String caseId = "1234567891234567L";
-    //        String eventId = "fl401RequestSupport";
-    //        String partyId = "e3ceb507-0137-43a9-8bd3-85dd23720648";
-    //        FlagDetailRequest flagDetailRequest = FlagDetailRequest.builder()
-    //            .name("Support filling in forms")
-    //            .name_cy("Cymorth i lenwi ffurflenni")
-    //            .hearingRelevant(YesOrNo.No)
-    //            .flagCode("RA0018")
-    //            .status("Requested")
-    //            .availableExternally(YesOrNo.Yes)
-    //            .build();
-    //        List<FlagDetailRequest> flagDetailsRequest = Collections.singletonList(flagDetailRequest);
-    //        CitizenPartyFlagsRequest partyRequestFlags = CitizenPartyFlagsRequest.builder()
-    //            .caseTypeOfApplication("FL401")
-    //            .partyIdamId(partyId)
-    //            .partyExternalFlags(FlagsRequest.builder()
-    //                                    .details(flagDetailsRequest).build()).build();
-    //
-    //        Mockito.when(authorisationService.authoriseUser(authToken)).thenReturn(Boolean.FALSE);
-    //        Mockito.when(authorisationService.authoriseService(servAuthToken)).thenReturn(Boolean.TRUE);
-    //
-    //        ResponseEntity<Object> updateResponse = caseController.updateCitizenRAflags(
-    //        partyRequestFlags, eventId, caseId, authToken, servAuthToken);
-    //
-    //        assertThat(updateResponse.getStatusCode()).isEqualTo(HttpStatus.UNAUTHORIZED);
-    //    }
 }
