+++ src/test/java/uk/gov/hmcts/reform/prl/services/citizen/CaseServiceTest.java
@@ -2,43 +2,28 @@ package uk.gov.hmcts.reform.prl.services.citizen;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import javassist.NotFoundException;
 import org.junit.Assert;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.ccd.client.model.CaseEventDetail;
-import uk.gov.hmcts.reform.ccd.client.model.EventRequestData;
 import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
 import uk.gov.hmcts.reform.idam.client.IdamClient;
 import uk.gov.hmcts.reform.idam.client.models.UserDetails;
 import uk.gov.hmcts.reform.prl.clients.ccd.CcdCoreCaseDataService;
 import uk.gov.hmcts.reform.prl.constants.PrlAppsConstants;
-import uk.gov.hmcts.reform.prl.enums.CaseEvent;
 import uk.gov.hmcts.reform.prl.enums.PartyEnum;
 import uk.gov.hmcts.reform.prl.enums.State;
-import uk.gov.hmcts.reform.prl.enums.YesOrNo;
-import uk.gov.hmcts.reform.prl.enums.caseflags.PartyRole;
 import uk.gov.hmcts.reform.prl.mapper.citizen.CaseDataMapper;
 import uk.gov.hmcts.reform.prl.models.Element;
 import uk.gov.hmcts.reform.prl.models.UpdateCaseData;
-import uk.gov.hmcts.reform.prl.models.caseflags.AllPartyFlags;
-import uk.gov.hmcts.reform.prl.models.caseflags.Flags;
-import uk.gov.hmcts.reform.prl.models.caseflags.flagdetails.FlagDetail;
-import uk.gov.hmcts.reform.prl.models.caseflags.request.CitizenPartyFlagsRequest;
-import uk.gov.hmcts.reform.prl.models.caseflags.request.FlagDetailRequest;
-import uk.gov.hmcts.reform.prl.models.caseflags.request.FlagsRequest;
 import uk.gov.hmcts.reform.prl.models.caseinvite.CaseInvite;
 import uk.gov.hmcts.reform.prl.models.complextypes.PartyDetails;
 import uk.gov.hmcts.reform.prl.models.complextypes.WithdrawApplication;
@@ -54,10 +39,7 @@ import uk.gov.hmcts.reform.prl.services.tab.alltabs.AllTabServiceImpl;
 import uk.gov.hmcts.reform.prl.utils.CaseDetailsConverter;
 import uk.gov.hmcts.reform.prl.utils.CaseUtils;
 
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -728,293 +710,5 @@ public class CaseServiceTest {
         String isValid = caseService.validateAccessCode(authToken, s2sToken, caseId, accessCode);
 
         assertEquals(INVALID, isValid);
-    }
-
-    @Test
-    public void testGetPartyCaseFlags() {
-        User user1 = User.builder().idamId("applicant-1").build();
-        User user2 = User.builder().idamId("respondent-1").build();
-        User user3 = User.builder().idamId("respondent-2").build();
-        PartyDetails applicant = PartyDetails.builder().user(user1).email("testappl@hmcts.net").firstName(
-            "Applicant 1 FN").lastName("Applicant 1 LN").build();
-        PartyDetails respondent1 = PartyDetails.builder().user(user2).email("testresp1@hmcts.net").firstName(
-            "Respondent 1 FN").lastName("Respondent 1 LN").build();
-        PartyDetails respondent2 = PartyDetails.builder().user(user3).email("testresp2@hmcts.net").firstName(
-            "Respondent 2 FN").lastName("Respondent 2 LN").build();
-
-        FlagDetail flagDetailRequestForFillingForms = FlagDetail.builder()
-            .name("Support filling in forms")
-            .name_cy("Cymorth i lenwi ffurflenni")
-            .hearingRelevant(No)
-            .flagCode("RA0018")
-            .status("Requested")
-            .availableExternally(Yes)
-            .build();
-        FlagDetail flagDetailRequestForHearing = FlagDetail.builder()
-            .name("Private waiting area")
-            .name_cy("Ystafell aros breifat")
-            .hearingRelevant(Yes)
-            .flagCode("RA0033")
-            .status("Requested")
-            .availableExternally(Yes)
-            .build();
-
-        Flags applicant1PartyFlags = Flags.builder().roleOnCase("Applicant 1").partyName("Applicant 1 FN Applicant 1 LN").details(
-            Collections.singletonList(element(flagDetailRequestForFillingForms))).build();
-        Flags respondent1PartyFlags = Flags.builder().roleOnCase("Respondent 1").partyName(
-            "Respondent 1 FN Respondent 1 LN").details(Collections.singletonList(element(flagDetailRequestForHearing))).build();
-        Flags respondent2PartyFlags = Flags.builder().roleOnCase("Respondent 2").partyName(
-            "Respondent 2 FN Respondent 2 LN").details(Arrays.asList(
-            element(flagDetailRequestForFillingForms),
-            element(flagDetailRequestForHearing)
-        )).build();
-
-        CaseData caseData = CaseData.builder()
-            .id(1234567891234567L)
-            .caseTypeOfApplication("C100")
-            .applicants(Collections.singletonList(element(applicant)))
-            .respondents(Arrays.asList(element(respondent1), element(respondent2)))
-            .allPartyFlags(AllPartyFlags.builder().caApplicant1ExternalFlags(applicant1PartyFlags).caRespondent1ExternalFlags(
-                respondent1PartyFlags).caRespondent2ExternalFlags(respondent2PartyFlags).build())
-            .build();
-        Map<String, Object> stringObjectMap = caseData.toMap(new ObjectMapper());
-        CaseDetails caseDetails = CaseDetails.builder()
-            .id(1234567891234567L)
-            .data(stringObjectMap)
-            .build();
-        when(caseService.getCase(authToken, caseId)).thenReturn(caseDetails);
-        when(objectMapper.convertValue(stringObjectMap, CaseData.class)).thenReturn(caseData);
-
-        // Unhappy path - when the request is valid, but party details is invalid.
-        Flags invalidUserExternalFlag = caseService.getPartyCaseFlags(authToken, caseId, "applicant-2");
-        Assert.assertNull(invalidUserExternalFlag);
-
-        // Happy path 1 - when the request is valid and respondent party external flags is retrieved from the existing case data.
-        when(partyLevelCaseFlagsService.getPartyCaseDataExternalField(
-            "C100",
-            PartyRole.Representing.CARESPONDENT,
-            1
-        )).thenReturn("caRespondent2ExternalFlags");
-        when(objectMapper.convertValue(Mockito.any(), Mockito.eq(Flags.class))).thenReturn(respondent2PartyFlags);
-        Flags respondentExternalFlag = caseService.getPartyCaseFlags(authToken, caseId, "respondent-2");
-        Assert.assertNotNull(respondentExternalFlag);
-        Assert.assertEquals("Respondent 2 FN Respondent 2 LN", respondentExternalFlag.getPartyName());
-
-        // Happy path 2 - when the request is valid and applicant party external flags is retrieved from the existing case data.
-        when(partyLevelCaseFlagsService.getPartyCaseDataExternalField(
-            "C100",
-            PartyRole.Representing.CAAPPLICANT,
-            0
-        )).thenReturn("caApplicant1ExternalFlags");
-        when(objectMapper.convertValue(Mockito.any(), Mockito.eq(Flags.class))).thenReturn(applicant1PartyFlags);
-        Flags applicantExternalFlag = caseService.getPartyCaseFlags(authToken, caseId, "applicant-1");
-        Assert.assertNotNull(applicantExternalFlag);
-        Assert.assertEquals(applicant1PartyFlags, applicantExternalFlag);
-    }
-
-    @Test
-    @Ignore
-    public void testUpdateCitizenRaflags() {
-        User user1 = User.builder().idamId("applicant-1").build();
-        User user2 = User.builder().idamId("respondent-1").build();
-        User user3 = User.builder().idamId("respondent-2").build();
-        PartyDetails applicant = PartyDetails.builder().user(user1).email("testappl@hmcts.net").firstName(
-            "Applicant 1 FN").lastName("Applicant 1 LN").build();
-        PartyDetails respondent1 = PartyDetails.builder().user(user2).email("testresp1@hmcts.net").firstName(
-            "Respondent 1 FN").lastName("Respondent 1 LN").build();
-        PartyDetails respondent2 = PartyDetails.builder().user(user3).email("testresp2@hmcts.net").firstName(
-            "Respondent 2 FN").lastName("Respondent 2 LN").build();
-
-        FlagDetail flagDetailRequestForFillingForms = FlagDetail.builder()
-            .name("Support filling in forms")
-            .name_cy("Cymorth i lenwi ffurflenni")
-            .hearingRelevant(No)
-            .flagCode("RA0018")
-            .status("Requested")
-            .dateTimeCreated(LocalDateTime.parse(
-                "2023-11-11T12:12:12.000Z",
-                DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
-            ))
-            .availableExternally(Yes)
-            .build();
-        FlagDetail flagDetailRequestForHearing = FlagDetail.builder()
-            .name("Private waiting area")
-            .name_cy("Ystafell aros breifat")
-            .hearingRelevant(Yes)
-            .flagCode("RA0033")
-            .status("Requested")
-            .dateTimeCreated(LocalDateTime.parse(
-                "2023-11-11T12:13:02.000Z",
-                DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
-            ))
-            .availableExternally(Yes)
-            .build();
-
-        FlagDetailRequest updateFlagDetailRequestForFillingForms = FlagDetailRequest.builder()
-            .name("Support filling in forms")
-            .name_cy("Cymorth i lenwi ffurflenni")
-            .hearingRelevant(YesOrNo.No)
-            .flagCode("RA0018")
-            .status("Inactive")
-            .dateTimeCreated(LocalDateTime.parse(
-                "2023-11-11T12:13:02.000Z",
-                DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
-            ))
-            .dateTimeModified(LocalDateTime.parse(
-                "2023-11-12T10:09:21.000Z",
-                DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
-            ))
-            .availableExternally(YesOrNo.Yes)
-            .build();
-
-        FlagDetailRequest updateFlagDetailRequestForHearing = FlagDetailRequest.builder()
-            .name("Private waiting area")
-            .name_cy("Ystafell aros breifat")
-            .hearingRelevant(Yes)
-            .flagCode("RA0033")
-            .status("Inactive")
-            .dateTimeCreated(LocalDateTime.parse(
-                "2023-11-11T12:13:02.000Z",
-                DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
-            ))
-            .dateTimeModified(LocalDateTime.parse(
-                "2023-11-12T10:09:21.000Z",
-                DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
-            ))
-            .availableExternally(Yes)
-            .build();
-
-        Flags applicant1PartyFlags = Flags.builder().roleOnCase("Applicant 1").partyName("Applicant 1 FN Applicant 1 LN").details(
-            Collections.singletonList(element(flagDetailRequestForFillingForms))).build();
-        Flags respondent1PartyFlags = Flags.builder().roleOnCase("Respondent 1").partyName(
-            "Respondent 1 FN Respondent 1 LN").details(Collections.singletonList(element(flagDetailRequestForHearing))).build();
-        Flags respondent2PartyFlags = Flags.builder().roleOnCase("Respondent 2").partyName(
-            "Respondent 2 FN Respondent 2 LN").details(Arrays.asList(
-            element(flagDetailRequestForFillingForms),
-            element(flagDetailRequestForHearing)
-        )).build();
-        CitizenPartyFlagsRequest updatePartyFlagsRequest;
-        updatePartyFlagsRequest = CitizenPartyFlagsRequest.builder()
-            .caseTypeOfApplication("C100")
-            .partyIdamId("respondent-1")
-            .partyExternalFlags(FlagsRequest.builder()
-                                    .details(Arrays.asList(
-                                        updateFlagDetailRequestForFillingForms,
-                                        updateFlagDetailRequestForHearing
-                                    )).build()).build();
-
-        CaseData caseData = CaseData.builder()
-            .id(1234567891234567L)
-            .caseTypeOfApplication("C100")
-            .applicants(Collections.singletonList(element(applicant)))
-            .respondents(Arrays.asList(element(respondent1), element(respondent2)))
-            .allPartyFlags(AllPartyFlags.builder().caApplicant1ExternalFlags(applicant1PartyFlags).caRespondent1ExternalFlags(
-                respondent1PartyFlags).caRespondent2ExternalFlags(respondent2PartyFlags).build())
-            .build();
-        ObjectMapper mapper = new ObjectMapper();
-        mapper.registerModule(new JavaTimeModule());
-        mapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
-        Map<String, Object> stringObjectMap = caseData.toMap(mapper);
-        CaseDetails caseDetails = CaseDetails.builder()
-            .id(1234567891234567L)
-            .data(stringObjectMap)
-            .build();
-        when(idamClient.getUserDetails(authToken)).thenReturn(userDetails);
-
-        when(coreCaseDataService.eventRequest(CaseEvent.fromValue("c100RequestSupport"), systemUserId)).thenReturn(
-            EventRequestData.builder().build());
-        StartEventResponse startEventResponse = StartEventResponse.builder().eventId("c100RequestSupport")
-            .caseDetails(caseDetails)
-            .token(eventToken).build();
-        when(coreCaseDataService.startUpdate(
-            authToken,
-            EventRequestData.builder().build(),
-            caseId,
-            false
-        )).thenReturn(
-            startEventResponse);
-        when(objectMapper.convertValue(Mockito.any(), Mockito.eq(CaseData.class))).thenReturn(caseData);
-
-        // Happy path 1 - when the request is valid and respondent party external flags is updated in the response.
-        when(objectMapper.convertValue(Mockito.any(), Mockito.eq(Flags.class))).thenReturn(respondent1PartyFlags);
-        when(partyLevelCaseFlagsService.getPartyCaseDataExternalField(
-            "C100",
-            PartyRole.Representing.CARESPONDENT,
-            0
-        )).thenReturn("caRespondent1ExternalFlags");
-        ResponseEntity<Object> updatePartyFlagsResponse;
-        updatePartyFlagsResponse = caseService.createCitizenReasonableAdjustmentsFlags(
-            caseId,
-            "c100RequestSupport",
-            authToken,
-            updatePartyFlagsRequest
-        );
-        assertThat(updatePartyFlagsResponse.getStatusCode()).isEqualTo(HttpStatus.OK);
-
-        // Happy path 2 - when the request is valid and applicant party external flags is updated in the response.
-        when(objectMapper.convertValue(Mockito.any(), Mockito.eq(Flags.class))).thenReturn(applicant1PartyFlags);
-        when(partyLevelCaseFlagsService.getPartyCaseDataExternalField(
-            "C100",
-            PartyRole.Representing.CAAPPLICANT,
-            0
-        )).thenReturn("caApplicant1ExternalFlags");
-        updatePartyFlagsRequest = CitizenPartyFlagsRequest.builder()
-            .caseTypeOfApplication("C100")
-            .partyIdamId("applicant-1")
-            .partyExternalFlags(FlagsRequest.builder()
-                                    .details(Arrays.asList(
-                                        updateFlagDetailRequestForFillingForms,
-                                        updateFlagDetailRequestForHearing
-                                    )).build()).build();
-        updatePartyFlagsResponse = caseService.createCitizenReasonableAdjustmentsFlags(
-            caseId,
-            "c100RequestSupport",
-            authToken,
-            updatePartyFlagsRequest
-        );
-        assertThat(updatePartyFlagsResponse.getStatusCode()).isEqualTo(HttpStatus.OK);
-
-        // Unhappy path 1 - when the request is not valid.
-        updatePartyFlagsRequest = CitizenPartyFlagsRequest.builder()
-            .caseTypeOfApplication("C100")
-            .partyIdamId("respondent-1").build();
-        updatePartyFlagsResponse = caseService.createCitizenReasonableAdjustmentsFlags(
-            caseId,
-            "c100RequestSupport",
-            authToken,
-            updatePartyFlagsRequest
-        );
-        assertThat(updatePartyFlagsResponse.getStatusCode()).isEqualTo(HttpStatus.BAD_REQUEST);
-
-        // Unhappy path 2 - when the request is valid, but party details is invalid.
-        updatePartyFlagsRequest = CitizenPartyFlagsRequest.builder()
-            .caseTypeOfApplication("C100")
-            .partyIdamId("respondent-3").partyExternalFlags(FlagsRequest.builder().build()).build();
-        updatePartyFlagsResponse = caseService.createCitizenReasonableAdjustmentsFlags(
-            caseId,
-            "c100RequestSupport",
-            authToken,
-            updatePartyFlagsRequest
-        );
-        assertThat(updatePartyFlagsResponse.getStatusCode()).isEqualTo(HttpStatus.NOT_FOUND);
-        assertThat(updatePartyFlagsResponse.getBody()).isEqualTo("party details not found");
-
-        // Unhappy path 3 - when the request is valid, but party external flag details not present in the case data.
-        when(partyLevelCaseFlagsService.getPartyCaseDataExternalField(
-            "C100",
-            PartyRole.Representing.CARESPONDENT,
-            6
-        )).thenReturn("caRespondent7ExternalFlags");
-        updatePartyFlagsRequest = CitizenPartyFlagsRequest.builder()
-            .caseTypeOfApplication("C100")
-            .partyIdamId("respondent-2").partyExternalFlags(FlagsRequest.builder().build()).build();
-        updatePartyFlagsResponse = caseService.createCitizenReasonableAdjustmentsFlags(
-            caseId,
-            "c100RequestSupport",
-            authToken,
-            updatePartyFlagsRequest
-        );
-        assertThat(updatePartyFlagsResponse.getStatusCode()).isEqualTo(HttpStatus.NOT_FOUND);
-        assertThat(updatePartyFlagsResponse.getBody()).isEqualTo("party external flag details not found");
     }
 }
